* master: (272 commits)
  GLUE-1623 update inverse condition
  PS-9779: ignored phpstan error
  GLUE-8286 AA review fix
  TE-3688 downgrad PHPStan level to 4 because of Parameter #2  of function array_filter expects callable(mixed,mixed): bool, 'strlen' given.
  TE-3688 Updated tests.
  TE-3688 Updated tests.
  TE-3688 Symfony Form must be below 4.
  TE-3688 Added compatibility for Symfony v3 and v4.
  TE-3688 Added required arguments for SplFileInfo.
  TE-3688 Updated symfony/console version constraint for Symfony module
  TE-3688 Updated Console.
  TE-3688 Updated Console to use new Terminal and cleaner formatting for messages.
  TE-3688 Fixed console tests.
  TE-3688 Fixed codeception.yml.
  TE-3688 Allow symfony components v4.
  CMS-2104: updated dependencies to `spryker/content`
  CMS-2104: updated dependencies to `spryker/content`
  cms-1864: updated composer.json of ContentBannerGui
  GLUE-1623 fixed condition for phpstan standalone test
  GLUE-1623 update composer.json
  ...

+++ Bundles/SalesQuantity/src/Spryker/Zed/SalesQuantity/SalesQuantityConfig.php
@@ -12,14 +12,14 @@ use Spryker\Zed\Kernel\AbstractBundleConfig;
 class SalesQuantityConfig extends AbstractBundleConfig
 {
     /**
-     * @const float|null Controls the threshold for quantity which above the quantity should not be splitted. Null value inactivates the threshold.
+     * @const int|null Controls the threshold for quantity which above the quantity should not be splitted. Null value inactivates the threshold.
      */
     protected const ITEM_NONSPLIT_QUANTITY_THRESHOLD = null;
 
     /**
-     * @return float|null
+     * @return int|null
      */
-    public function findItemQuantityThreshold(): ?float
+    public function findItemQuantityThreshold(): ?int
     {
         return static::ITEM_NONSPLIT_QUANTITY_THRESHOLD;
     }
